VOL-1334: Update linting and indentation of all contracts

+++ contracts/interfaces/ITransferManager.sol
@@ -5,9 +5,8 @@ pragma solidity =0.8.12;
 import "../libs/LibDeal.sol";
 
 abstract contract ITransferManager {
-
-    function _doTransfers(
-        LibDeal.DealSide memory left,
-        LibDeal.DealSide memory right
-    ) internal virtual returns (uint totalMakeValue, uint totalTakeValue);
+    function _doTransfers(LibDeal.DealSide memory left, LibDeal.DealSide memory right)
+        internal
+        virtual
+        returns (uint256 totalMakeValue, uint256 totalTakeValue);
 }
